Refactor send attendee invite email job, mail and template
