+++ resources/views/Emails/AttendeeInvite.blade.php
